inject config into cli
